backend/functions/requirements.txt
firebase_functions
plaid-python
openai